# services/identity-service/src/main/java/com/bk/arenax/identity/controller/dto/PasswordResetRequest.java
package com.bk.arenax.identity.controller.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record PasswordResetRequest(
        @NotBlank @Email @Size(max = 320) String email) {

    public PasswordResetRequest {
        email = email == null ? null : email.trim();
    }
}